Renamed lib to match upcoming changes in another branch.
Issue #11
