+++ lib/screenshot.js
